Crate monitord

monitord Crate

monitord is a library to gather statistics about systemd.

Modules

config

json

json module

logging

machines

networkd

networkd module

pid1

pid1 module

system

system module

units

units module

Structs

MachineStats

MonitordStats

Main monitord stats struct collection all enabled stats

Constants

DEFAULT_DBUS_ADDRESS

Functions

print_stats

Print statistics in the format set in configuration

stat_collector

Main statictic collection function running what’s required by configuration in parallel Takes an optional locked stats struct to update and to output stats to STDOUT or not