monitord::varlink::network

Trait NetworkChain

pub trait NetworkChain<'c, S>
where
    S: Socket,
{
    // Required method
    fn describe(self) -> Result<Chain<'c, S>>;
}

Extension trait for adding proxy calls to any chain.

This trait provides methods to add proxy calls to a chain of method calls.

Required Methods

fn describe(self) -> Result<Chain<'c, S>>

Add a #method_name call to this chain.