Crate monitord

monitord Crate

monitord is a library to gather statistics about systemd. Some APIs are a little ugly due to being a configparser INI based configuration driven CLL at heart.

Modules

json

json module

logging

networkd

networkd module

pid1

pid1 module

system

system module

units

units module

Structs

MonitordStats

Main monitord stats struct collection all enabled stats

Constants

DEFAULT_DBUS_ADDRESS

Functions

print_stats

Print statistics in the format set in configuration

stat_collector

Main statictic collection function running what’s required by configuration