monitord

Module dbus_stats

dbus_stats module

Handle getting statistics of our Dbus daemon/broker

Structs

CurMaxPair

DBusBrokerCGroupAccounting

DBusBrokerPeerAccounting

DBusBrokerUserAccounting

DBusStats

Functions

parse_dbus_stats

Pull all units from dbus and count how system is setup and behaving

update_dbus_stats

Async wrapper than can update dbus stats when passed a locked struct