monitord
Module networkd

networkd module

All structs, enums and methods specific to systemd-networkd. Enumerations were copied from https://github.com/systemd/systemd/blob/main/src/libsystemd/sd-network/network-util.h

Structs

AddressStateIter

An iterator over the variants of AddressState

AdminStateIter

An iterator over the variants of AdminState

BoolStateIter

An iterator over the variants of BoolState

CarrierStateIter

An iterator over the variants of CarrierState

InterfaceState

Main per interface networkd state structure

NetworkdState

Main networkd structure with per interface state and a count of managed interfaces

OnlineStateIter

An iterator over the variants of OnlineState

OperStateIter

An iterator over the variants of OperState

Enums

AddressState

Enumeration of networkd address states

AdminState

Enumeration of interface administratve states

BoolState

Enumeration of a true (yes) / false (no) options - e.g. required for online

CarrierState

Enumeration of networkd physical signal / state of interfaces

OnlineState

Enumeration of the networkd online state

OperState

Enumeration of networkd’s operational state

Constants

NETWORKD_STATE_FILES

Functions

parse_interface_state_files

Parse interface state files in directory supplied

parse_interface_stats

Parse a networkd state file contents + convert int ID to name via DBUS

update_networkd_stats

Async wrapper than can update networkd stats when passed a locked struct